+++ src/clj_kondo/impl/config.clj
@@ -247,14 +247,28 @@
       (assoc cfg :linters linters))
     cfg))
 
+(defn- merge-fn-types [a b]
+  (deep-merge (update a :arities merge (:arities b))
+              (dissoc b :arities)))
+
+(defn- merge-type-mismatch-namespaces [a b]
+  (merge-with (partial merge-with merge-fn-types) a b))
+
 (defn merge-config!
   ([])
   ([cfg] cfg)
   ([cfg* cfg]
    (if (empty? cfg) cfg*
-       (let [cfg (cond-> cfg
+       (let [cfg* (cond-> cfg*
+                    (contains? (:linters cfg) :type-mismatch)
+                    (update-in [:linters :type-mismatch :namespaces]
+                               merge-type-mismatch-namespaces
+                               (get-in cfg [:linters :type-mismatch :namespaces])))
+             cfg (cond-> cfg
                    (contains? (:linters cfg) :if)
-                   (assoc-in [:linters :missing-else-branch] (:if (:linters cfg))))]
+                   (assoc-in [:linters :missing-else-branch] (:if (:linters cfg)))
+                   (contains? (:linters cfg) :type-mismatch)
+                   (update-in [:linters :type-mismatch] dissoc :namespaces))]
          (deep-merge cfg* cfg))))
   ([cfg* cfg & cfgs]
    (reduce merge-config! cfg* (cons cfg cfgs))))